Ensure that database is flushed periodically.
Added a timer to flush the database every 5 seconds in simserver. When we run the simulation
with no human agents, the database can't keep up with all of the logs, so we need to wait for it
periodically. (When we have human agents, the database will flush whenever there's downtime)